Implemented multiple target grovers as well as a function to plot for different ts
Not the final version, but can do multiple known targets in know array size. Also there is a plot for probability vs t but this has some errors with taking the wrong probability that I will sort out later.
Next step: make it work for array size that are not powers of 2 and unknown number of solutions. xx